UserServiceTest -> UserCommandServiceTest 유저 탈퇴 관련 테스트 이동
